+++ RiotGamesAPIWrapper/HechTex.RiotGamesAPIWrapper/Cache/RunePageCache/DefaultRunePageCache.cs
@@ -22,7 +22,7 @@ namespace HechTex.RiotGamesAPIWrapper.Cache.RunePageCache
         /// queries/calls to the API itself.</param>
         /// <param name="region">The region.</param>
         /// <param name="summonerId">The summoner's id.</param>
-        internal DefaultRunePageCache(APICaller apiCaller, Regions region, int summonerId)
+        internal DefaultRunePageCache(APICaller apiCaller, Regions region, long summonerId)
             : base(apiCaller, region, summonerId.ToString())
         {
             CacheMethod = CacheMethod.Default;
